// proj-main/src/types.ts
export interface Abbreviation {
  usTerm: string;
  ukTerm: string;
  name: string;
  description: string;
  symbol?: string;
  difficulty: 'Basic' | 'Intermediate' | 'Advanced';
}